% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castaway_convert.R
\name{castaway_convert}
\alias{castaway_convert}
\title{Castaway CTD to netCDF conversion}
\usage{
castaway_convert(files, destination)
}
\arguments{
\item{files}{A character vector of a directory in which castaway CTD CSV
files are stored}

\item{destination}{an existing folder to save netCDF files to}
}
\value{
\code{castaway_convert()} creates netCDF files from castaway CTD CSV files
}
\description{
Castaway CTD to netCDF conversion
}
\examples{
castaway_convert("data/")

}
