% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castaway_convert.R
\name{castaway_convert}
\alias{castaway_convert}
\title{Castaway CTD to netCDF conversion}
\usage{
castaway_convert(
  files,
  destination = "R:/Science/CESD/CESD_DataManagement/data_out/castawayCTD",
  level = "raw",
  project_lead,
  project_name
)
}
\arguments{
\item{files}{A character vector of a directory in which castaway CTD CSV
files are stored.}

\item{destination}{an existing folder to save netCDF files to}

\item{level}{either \code{raw} or \code{processed} based on the level of
processing completed using the castaway CTD software}

\item{project_lead}{an optional field to add the project lead's name to the file global attributes}

\item{project_name}{an optional field to add a project name to the file global attributes}
}
\value{
\code{castaway_convert()} creates netCDF files from castaway CTD CSV
  files
}
\description{
Castaway CTD to netCDF conversion
}
\examples{
castaway_convert("data/")

}
