% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castaway_convert.R
\name{castaway_convert}
\alias{castaway_convert}
\title{Castaway CTD to netCDF conversion}
\usage{
castaway_convert(
  files,
  destination = "R:/Science/CESD/CESD_DataManagement/data_out/castawayCTD",
  level = "raw"
)
}
\arguments{
\item{files}{A character vector of a directory in which castaway CTD CSV
files are stored. Must be in "Raw" format from the Castaway CTD Software
(only variable for \code{Time (Seconds)},    \code{Pressure
(Decibar)},    \code{Temperature (Celsius)}, and \code{Conductivity
(MicroSiemens per Centimeter)})}

\item{destination}{an existing folder to save netCDF files to}

\item{level}{either \code{raw} or \code{processed} based on the level of
processing completed using the castaway CTD software}
}
\value{
\code{castaway_convert()} creates netCDF files from castaway CTD CSV
  files
}
\description{
Castaway CTD to netCDF conversion
}
\examples{
castaway_convert("data/")

}
